- VS2010 warning fixed
git-svn-id: svn://rdo.rk9.bmstu.ru/var/svn/root@12999 23f926e6-80f1-49ca-85a0-60d4a27bb3ac

+++ rdo/packages/compiler/parser/trunk/rdoparser_lexer.cpp
@@ -65,7 +65,7 @@ int RDOLexer::LexerInput(PTR(char) buf, int max_size)
 	}
 	else
 	{
-		return m_yyin->gcount();
+		return (int)m_yyin->gcount();
 	}
 #endif
 }
